documentation cleanups, typo fixes, improvements in user-visible strings

+++ src/play/gesture/PanGestureController.m
@@ -290,7 +290,7 @@
     }
     default:
     {
-      DDLogDebug(@"handlePanFrom, unhandled recognizerState = %ld", recognizerState);
+      DDLogDebug(@"handlePanFrom, unhandled recognizerState = %ld", (long)recognizerState);
 
       break;
     }
